Fix float/int error

+++ project_04/map.h
@@ -5,7 +5,7 @@
 
 namespace uno {
 
-constexpr array<int, 3> map(const array<float, 3>& sensorValues) noexcept {
+constexpr array<int, 3> map(const array<int, 3>& sensorValues) noexcept {
   return {static_cast<int>(sensorValues[0]/4),
           static_cast<int>(sensorValues[1]/4),
           static_cast<int>(sensorValues[2]/4)};
