+++ project_04/read.cpp
@@ -6,8 +6,8 @@
 
 namespace uno {
 
-array<float, 3> read(const array<AnalogInput, 3>& photoresisters) noexcept {
-  array<float, 3> rgb;
+array<int, 3> read(const array<AnalogInput, 3>& photoresisters) noexcept {
+  array<int, 3> rgb;
   transform(photoresisters.begin(), photoresisters.end(),
             rgb.begin(),
             [](const AnalogInput& in) { return in.read(); });
